pyproject.toml
[project]
name = "beter_woorden"
version = "0.1.0"
description = "Dutch practice"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "langchain>=1.1.2",
    "langchain-google-genai>=3.2.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "pydantic>=2.12.5",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "uvicorn>=0.38.0",
]

[tool.uv.workspace]
members = [
    "worker",
]
